requirements.txt
amqp==1.4.9
anyjson==0.3.3
argon2-cffi==19.2.0
attrs==19.2.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
bleach==3.1.0
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.10
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-classy-tags==1.0.0
django-cms==3.7.1
django-filter==2.3.0
django-formtools==2.2
django-js-asset==1.2.2
django-markdown==0.8.4
django-rest-framework==0.1.0
django-sekizai==1.1.0
django-timezone-field==4.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
Faker==4.0.0
html5lib==1.0.1
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
kombu==3.0.37
lxml==4.4.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mysql-connector-python==8.0.18
mysqlclient==1.4.6
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.18.1
oauthlib==3.1.0
openpyxl==3.0.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.2
pycparser==2.20
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.6
pyrsistent==0.15.4
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
redis==3.5.2
requests==2.22.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.1.0
